+++ node-server/bin/app.js
@@ -26,7 +26,6 @@ import sessionSequelize from "connect-session-sequelize";
 import indexRouter from "../routes/index.js";
 import userRouter from "../routes/user.js";
 import communityRouter from "../routes/community.js";
-import kakaoRouter from "../routes/kakaoAPI.js";
 import createrAPI from "../routes/createrAPI.js";
 // create express framework
 const app = express();
