+++ node-server/config/db_config.js
@@ -6,6 +6,7 @@ export default {
     host: "127.0.0.1",
     dialect: "mysql",
     logging: true,
+    port: "3307",
     timezone: "+09:00",
   },
   test: {
