+++ react-client/src/config/kakao_config_sample.js
@@ -1,3 +1,4 @@
 export const KAKAO_APP_ADMIN_KEY = "";
 
 export const READY_URL = "https://kapi.kakao.com/v1/payment/ready";
+export const APPROVE_URL = "https://kapi.kakao.com/v1/payment/approve";
