+++ react-client/src/nav/Navigation.js
@@ -4,6 +4,7 @@ import MainPage from "../comp/mainpage/MainPage";
 import MyPageMain from "../comp/mypage/MyPageMain";
 import Join from "../comp/login/Join";
 import Login from "../comp/login/Login";
+import Approve from "../comp/purchase/Approve";
 const router = createBrowserRouter([
   {
     path: "/",
@@ -13,6 +14,7 @@ const router = createBrowserRouter([
       { path: "/user", element: <Join /> },
       { path: "/user/login", element: <Login /> },
       { path: "/mypage", element: <MyPageMain /> },
+      { path: "/approval", element: <Approve /> },
     ],
   },
 ]);
